+++ src/main/java/io/redstudioragnarok/valkyrie/asm/ValkyriePlugin.java
@@ -2,7 +2,6 @@ package io.redstudioragnarok.valkyrie.asm;
 
 import dev.redstudio.redcore.utils.OptiNotFine;
 import io.redstudioragnarok.valkyrie.config.ValkyrieConfig;
-import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.relauncher.IFMLLoadingPlugin;
 import zone.rong.mixinbooter.IEarlyMixinLoader;